🔨refactor: dropdownとdialogの共通化
