+++ lib/login/login_model.dart
@@ -1,6 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import '../common/dialog.dart';
 import '../const/const.dart';
 
 final emailProvider =
@@ -45,33 +46,6 @@ Future<void> login(String email, String password) async {
   await auth.signInWithEmailAndPassword(
     email: email,
     password: password,
-  );
-}
-
-// ダイアログ
-Future<void> dialog(
-  BuildContext context,
-  String message, {
-  String btnText = 'OK',
-}) {
-  return showDialog(
-    barrierDismissible: false,
-    context: context,
-    builder: (BuildContext context) {
-      return WillPopScope(
-        // 戻るボタンを無効にする
-        onWillPop: () async => false,
-        child: AlertDialog(
-          title: Text(message),
-          actions: <Widget>[
-            TextButton(
-              child: Text(btnText),
-              onPressed: () => Navigator.of(context).pop(),
-            ),
-          ],
-        ),
-      );
-    },
   );
 }
