+++ lib/login/new.dart
@@ -2,6 +2,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:habitual_records/setup/setup.dart';
+import '../common/dialog.dart';
 import 'login_model.dart';
 
 class New extends StatelessWidget {
